requirements.txt
# Requirements for EvoSimC project
docker>=6.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.0.0
google-generativeai>=0.3.0
